+++ apollo.config.js
@@ -2,9 +2,9 @@ module.exports = {
   client: {
     name: "Wobbly",
     service: {
-      name: "Wobbly Staging",
-      url: "http://develop.wobbly.app",
-      skipSSLValidation: true
+      name: "Wobbly - develop",
+      url: "https://develop.wobbly.app",
+      skipSSLValidation: false
     }
   }
 };