+++ src/App.tsx
@@ -1,42 +1,33 @@
 import { ActionSheetProvider } from "@expo/react-native-action-sheet";
-import { Font, SecureStore } from "expo";
-import jwtDecode from "jwt-decode";
 import * as React from "react";
 import { ApolloProvider } from "react-apollo";
 import { NavigationContainerComponent } from "react-navigation";
-import SentryExpo from "sentry-expo";
 
 import { client, initClient } from "./apolloClient";
 import AppNavigation from "./AppNavigation";
 import AppWithSubscriptions from "./AppWithSubscriptions";
-import { SplashScreen } from "./components/screens";
-import { config } from "./config";
-import { fonts } from "./fonts";
 import { OWN_INFO_QUERY, OwnInfoQuery } from "./graphql/queries";
 import { NavigationService } from "./services";
 
 interface IAppState {
   clientHasLoaded: boolean;
-  fontsHaveLoaded: boolean;
 }
 export default class App extends React.Component<{}, IAppState> {
   public constructor(props: {}) {
     super(props);
-    this.state = { clientHasLoaded: false, fontsHaveLoaded: false };
+    this.state = { clientHasLoaded: false };
   }
 
   public async componentDidMount() {
-    await this.initSentry();
-    await this.initClient();
-    await this.initFonts();
+    await initClient();
+    this.setState({ clientHasLoaded: true });
   }
 
   public render() {
-    if (!this.state.clientHasLoaded || !this.state.fontsHaveLoaded || !client) {
-      return <SplashScreen />;
+    if (!this.state.clientHasLoaded) {
+      return null;
     }
-    const navigationPersistenceKey = __DEV__ ? "NavigationStateDEV" : null;
-
+    const navigationPersistenceKey = __DEV__ ? null : null; // TODO @tao
     return (
       <ApolloProvider client={client}>
         <OwnInfoQuery query={OWN_INFO_QUERY}>
@@ -55,29 +46,5 @@ export default class App extends React.Component<{}, IAppState> {
         </OwnInfoQuery>
       </ApolloProvider>
     );
-  }
-
-  private async initSentry() {
-    if (!config.sentryDsn) {
-      return;
-    }
-    // SentryExpo.enableInExpoDevelopment = true;
-    SentryExpo.config(config.sentryDsn).install();
-    const authToken = await SecureStore.getItemAsync("token");
-    if (!authToken) {
-      return;
-    }
-    const decodedJwt = jwtDecode<{ personId: string }>(authToken);
-    SentryExpo.setUserContext({ id: decodedJwt.personId });
-  }
-
-  private async initClient() {
-    await initClient();
-    this.setState({ clientHasLoaded: true });
-  }
-
-  private async initFonts() {
-    await Font.loadAsync(fonts);
-    this.setState({ fontsHaveLoaded: true });
   }
 }