+++ src/components/molecules/NonIdealState.tsx
@@ -25,7 +25,8 @@ const style = StyleSheet.create({
   container: {
     flex: 1,
     alignItems: "center",
-    justifyContent: "center"
+    justifyContent: "center",
+    backgroundColor: "white"
   },
   title: {
     marginVertical: 5,