+++ src/components/organisms/LoadingState.tsx
@@ -11,7 +11,8 @@ const style = StyleSheet.create({
   container: {
     flex: 1,
     justifyContent: "center",
-    alignItems: "center"
+    alignItems: "center",
+    backgroundColor: "white"
   }
 });