+++ src/components/screens/GroupDetailsScreen.stories.tsx
@@ -1,11 +1,7 @@
 import { storiesOf } from "@storybook/react";
 
 import { screenStory } from "../../__stories__/storyWrapper";
-import { someId } from "../../__stories__/testData";
 
 import GroupDetailsScreen from "./GroupDetailsScreen";
 
-storiesOf("screens/GroupDetailsScreen", module).add(
-  "Standard",
-  screenStory(GroupDetailsScreen, { navigationParams: { groupId: someId() } })
-);
+storiesOf("screens/GroupDetailsScreen", module).add("Standard", screenStory(GroupDetailsScreen));