+++ src/graphql/queries/types.ts
@@ -1,5 +1,6 @@
 import { Query, QueryResult } from "react-apollo";
 
+import { getCurrentGroupId } from "../../generated/getCurrentGroupId";
 import { getGroupDetails } from "../../generated/getGroupDetails";
 import { getGroups } from "../../generated/getGroups";
 import { getOwnInfo } from "../../generated/getOwnInfo";
@@ -14,6 +15,9 @@ export type OwnInfoQueryResult = QueryResult<getOwnInfo>;
 
 export class GroupsQuery extends Query<getGroups> {}
 export type GroupsQueryResult = QueryResult<getGroups>;
+
+export class CurrentGroupIdQuery extends Query<getCurrentGroupId> {}
+export type CurrentGroupIdQueryResult = QueryResult<getCurrentGroupId>;
 
 export class GroupDetailsQuery extends Query<getGroupDetails> {}
 export type GroupDetailsQueryResult = QueryResult<getGroupDetails>;