+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add [danger](https://danger.systems/js/)
 - Set up error reporting with Sentry
 - Set up push notifications
+- New drawer navigation for the list of groups
 
 ## [0.0.1-alpha.1] - 2019-04-11
 ### Added